Fixed Joke
Responsiveness of all devices is still a concern for the portfolio. It works on screen sizes similar to the iphone SE just fine but anything other than that is a bit of a 50/50 crapshoot on whether or not the image will be aligned if there is something else not being accounted for that will break.