New additons
Added smooth scroll with the aid of JQuery.
Began adding projects section. 
Images of project and direct links have been added. Next step will be to create a scroll up effect on card information. Perhaps even a way for searchers to find more of my work if they want to.